pyproject.toml
[tool.poetry]
name = "discord-bot-template"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.8"
discord = "^1.0.1"
flask = "^1.1.2"
psutil = "^5.8.0"
asset = "^0.6.13"
colour = "^0.1.5"
enums = "^0.0.2"
flags = "^0.0.1"
ping3 = "^2.6.6"
utils = "^1.0.1"
conv = "^0.2"
django-encrypt = "^0.1"
dhooks = "^1.1.4"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
